+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/db/dao/impl/UserDataUserDAOJdbc.java
@@ -1,0 +1,47 @@
+package guru.qa.niffler.db.dao.impl;
+
+import guru.qa.niffler.db.ServiceDB;
+import guru.qa.niffler.db.dao.UserDataUserDAO;
+import guru.qa.niffler.db.jdbc.DataSourceProvider;
+import guru.qa.niffler.db.model.userdata.UserDataUserEntity;
+import guru.qa.niffler.model.CurrencyValues;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public class UserDataUserDAOJdbc implements UserDataUserDAO {
+
+    private static final DataSource userdataDs = DataSourceProvider.INSTANCE.getDataSource(ServiceDB.USERDATA);
+
+    @Override
+    public int createUserInUserData(UserDataUserEntity user) {
+        int createdRows;
+        try (Connection conn = userdataDs.getConnection();
+             PreparedStatement usersPs = conn.prepareStatement("INSERT INTO users (username, currency) VALUES (?, ?)")) {
+
+            usersPs.setString(1, user.getUsername());
+            usersPs.setString(2, CurrencyValues.RUB.name());
+
+            createdRows = usersPs.executeUpdate();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+        return createdRows;
+    }
+
+    @Override
+    public void deleteUserByNameInUserData(String username) {
+        try (Connection conn = userdataDs.getConnection()) {
+            try (PreparedStatement usersPs = conn.prepareStatement(
+                    "DELETE FROM users WHERE username = ?")) {
+
+                usersPs.setString(1, username);
+                usersPs.executeUpdate();
+            }
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
