+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/DBUser.java
@@ -1,6 +1,6 @@
 package guru.qa.niffler.jupiter.annotation;
 
-import guru.qa.niffler.jupiter.extension.UserExtension;
+import guru.qa.niffler.jupiter.extension.DBUserExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.lang.annotation.ElementType;
@@ -10,6 +10,6 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@ExtendWith(UserExtension.class)
+@ExtendWith(DBUserExtension.class)
 public @interface DBUser {
 }
